Sample multivariate normal distribution
A code to draw samples from a multivariate normal distribution by
computing the Cholesky Decomposition of the covariance matrix.
The covariance matrix is computed using the squared exponential kernel
from an n vector.
The n vector is plotted against the sample drawn from the Gaussian for
different values of s & l of the exponential kernel.